requirements.txt
# Hey Spider Robot - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
Flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Object Detection
ultralytics==8.0.228
torch==2.1.1
torchvision==0.16.1

# Raspberry Pi Hardware
adafruit-circuitpython-servokit==1.3.16
adafruit-circuitpython-ssd1306==1.9.9
RPi.GPIO==0.7.0
board==1.0.0
busio==1.9.1

# Picamera Support (choose one based on your system)
picamera2>=0.0.1  # Modern Raspberry Pi (Bullseye+)
# OR for older systems:
# picamera==1.13

# Speech Recognition
SpeechRecognition==3.10.0
pyaudio==0.2.13

# AI/ML
openai==1.3.6
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Optional Performance Tools
psutil==5.9.6  # System monitoring
tqdm==4.66.1   # Progress bars

# Development (optional)
pytest==7.4.3
black==23.12.0
flake8==6.1.0